Move some tests around to put them in right place
