+++ internal/controller/mcpserver_controller_storage_test.go
@@ -29,7 +29,6 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
@@ -1711,652 +1710,6 @@ var _ = Describe("MCPServer Controller - Storage Mounts", func() {
 			Expect(stderrors.As(err, &validationErr)).To(BeTrue())
 			Expect(validationErr.Reason).To(Equal(ReasonInvalid))
 			Expect(validationErr.Message).To(ContainSubstring("Invalid ConfigMap"))
-		})
-	})
-
-	Context("When reconciling a resource with resources", func() {
-		const resourceName = "test-resource-resources"
-
-		ctx := context.Background()
-
-		typeNamespacedName := types.NamespacedName{
-			Name:      resourceName,
-			Namespace: "default",
-		}
-
-		BeforeEach(func() {
-			mcpServer := newTestMCPServer(resourceName)
-			mcpServer.Spec.Runtime.Resources = &corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("100m"),
-					corev1.ResourceMemory: resource.MustParse("256Mi"),
-				},
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("500m"),
-					corev1.ResourceMemory: resource.MustParse("512Mi"),
-				},
-			}
-			Expect(k8sClient.Create(ctx, mcpServer)).To(Succeed())
-		})
-
-		AfterEach(func() {
-			resource := &mcpv1alpha1.MCPServer{}
-			err := k8sClient.Get(ctx, typeNamespacedName, resource)
-			if err == nil {
-				Expect(k8sClient.Delete(ctx, resource)).To(Succeed())
-			}
-		})
-
-		It("should create deployment with resources", func() {
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers).To(HaveLen(1))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("100m")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("256Mi")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("500m")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("512Mi")))
-			// Verify replicas defaults to 1 even when runtime is specified with other fields
-			Expect(*deployment.Spec.Replicas).To(Equal(int32(1)))
-		})
-
-		It("should update deployment when resources change", func() {
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			By("Reconciling to create the initial deployment with resources")
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("100m")))
-
-			By("Updating resources")
-			mcpServer := &mcpv1alpha1.MCPServer{}
-			Expect(k8sClient.Get(ctx, typeNamespacedName, mcpServer)).To(Succeed())
-			mcpServer.Spec.Runtime.Resources = &corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("200m"),
-					corev1.ResourceMemory: resource.MustParse("512Mi"),
-				},
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("1"),
-					corev1.ResourceMemory: resource.MustParse("1Gi"),
-				},
-			}
-			Expect(k8sClient.Update(ctx, mcpServer)).To(Succeed())
-
-			By("Reconciling again to pick up the change")
-			_, err = controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("200m")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("512Mi")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("1")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("1Gi")))
-		})
-
-		It("should update deployment when resources are removed", func() {
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			By("Reconciling to create the initial deployment with resources")
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("100m")))
-
-			By("Removing resources")
-			mcpServer := &mcpv1alpha1.MCPServer{}
-			Expect(k8sClient.Get(ctx, typeNamespacedName, mcpServer)).To(Succeed())
-			mcpServer.Spec.Runtime.Resources = nil
-			Expect(k8sClient.Update(ctx, mcpServer)).To(Succeed())
-
-			By("Reconciling again to pick up the change")
-			_, err = controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(BeEmpty())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(BeEmpty())
-		})
-
-		It("should handle resources with only requests (no limits)", func() {
-			mcpServer := newTestMCPServer("test-only-requests")
-			mcpServer.Spec.Runtime.Resources = &corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("100m"),
-					corev1.ResourceMemory: resource.MustParse("128Mi"),
-				},
-			}
-			Expect(k8sClient.Create(ctx, mcpServer)).To(Succeed())
-			defer func() {
-				err := k8sClient.Get(ctx, client.ObjectKey{Name: "test-only-requests", Namespace: "default"}, mcpServer)
-				if err == nil {
-					Expect(k8sClient.Delete(ctx, mcpServer)).To(Succeed())
-				}
-			}()
-
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: "test-only-requests", Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      "test-only-requests",
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("100m")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("128Mi")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(BeEmpty())
-		})
-
-		It("should handle resources with only limits (no requests)", func() {
-			mcpServer := newTestMCPServer("test-only-limits")
-			mcpServer.Spec.Runtime.Resources = &corev1.ResourceRequirements{
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("500m"),
-					corev1.ResourceMemory: resource.MustParse("512Mi"),
-				},
-			}
-			Expect(k8sClient.Create(ctx, mcpServer)).To(Succeed())
-			defer func() {
-				err := k8sClient.Get(ctx, client.ObjectKey{Name: "test-only-limits", Namespace: "default"}, mcpServer)
-				if err == nil {
-					Expect(k8sClient.Delete(ctx, mcpServer)).To(Succeed())
-				}
-			}()
-
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: "test-only-limits", Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      "test-only-limits",
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(BeEmpty())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("500m")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("512Mi")))
-		})
-
-		It("should handle resources with only CPU (no memory)", func() {
-			mcpServer := newTestMCPServer("test-only-cpu")
-			mcpServer.Spec.Runtime.Resources = &corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU: resource.MustParse("100m"),
-				},
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU: resource.MustParse("200m"),
-				},
-			}
-			Expect(k8sClient.Create(ctx, mcpServer)).To(Succeed())
-			defer func() {
-				err := k8sClient.Get(ctx, client.ObjectKey{Name: "test-only-cpu", Namespace: "default"}, mcpServer)
-				if err == nil {
-					Expect(k8sClient.Delete(ctx, mcpServer)).To(Succeed())
-				}
-			}()
-
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: "test-only-cpu", Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      "test-only-cpu",
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("100m")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).NotTo(HaveKey(corev1.ResourceMemory))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(HaveKeyWithValue(corev1.ResourceCPU, resource.MustParse("200m")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).NotTo(HaveKey(corev1.ResourceMemory))
-		})
-
-		It("should handle resources with only memory (no CPU)", func() {
-			mcpServer := newTestMCPServer("test-only-memory")
-			mcpServer.Spec.Runtime.Resources = &corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceMemory: resource.MustParse("256Mi"),
-				},
-				Limits: corev1.ResourceList{
-					corev1.ResourceMemory: resource.MustParse("512Mi"),
-				},
-			}
-			Expect(k8sClient.Create(ctx, mcpServer)).To(Succeed())
-			defer func() {
-				err := k8sClient.Get(ctx, client.ObjectKey{Name: "test-only-memory", Namespace: "default"}, mcpServer)
-				if err == nil {
-					Expect(k8sClient.Delete(ctx, mcpServer)).To(Succeed())
-				}
-			}()
-
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: "test-only-memory", Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      "test-only-memory",
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).NotTo(HaveKey(corev1.ResourceCPU))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Requests).To(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("256Mi")))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).NotTo(HaveKey(corev1.ResourceCPU))
-			Expect(deployment.Spec.Template.Spec.Containers[0].Resources.Limits).To(HaveKeyWithValue(corev1.ResourceMemory, resource.MustParse("512Mi")))
-		})
-	})
-
-	Context("Server-side apply for status updates", func() {
-		const resourceName = "test-ssa-status"
-		const subResourceStatus = "status"
-
-		ctx := context.Background()
-
-		typeNamespacedName := types.NamespacedName{
-			Name:      resourceName,
-			Namespace: "default",
-		}
-
-		BeforeEach(func() {
-			resource := newTestMCPServer(resourceName)
-			Expect(k8sClient.Create(ctx, resource)).To(Succeed())
-		})
-
-		AfterEach(func() {
-			resource := &mcpv1alpha1.MCPServer{}
-			err := k8sClient.Get(ctx, typeNamespacedName, resource)
-			if err == nil {
-				Expect(k8sClient.Delete(ctx, resource)).To(Succeed())
-			}
-		})
-
-		It("should use SubResourceApply for all status updates and never SubResourceUpdate or SubResourcePatch", func() {
-			applyCallCount := 0
-			updateCalled := false
-			patchCalled := false
-
-			wrappedClient, err := client.NewWithWatch(cfg, client.Options{Scheme: k8sClient.Scheme()})
-			Expect(err).NotTo(HaveOccurred())
-
-			interceptedClient := interceptor.NewClient(wrappedClient, interceptor.Funcs{
-				SubResourceApply: func(ctx context.Context, c client.Client, subResourceName string, obj runtime.ApplyConfiguration, opts ...client.SubResourceApplyOption) error {
-					if subResourceName == subResourceStatus {
-						applyCallCount++
-					}
-					return c.SubResource(subResourceName).Apply(ctx, obj, opts...)
-				},
-				SubResourceUpdate: func(ctx context.Context, c client.Client, subResourceName string, obj client.Object, opts ...client.SubResourceUpdateOption) error {
-					if subResourceName == subResourceStatus {
-						updateCalled = true
-					}
-					return c.SubResource(subResourceName).Update(ctx, obj, opts...)
-				},
-				SubResourcePatch: func(ctx context.Context, c client.Client, subResourceName string, obj client.Object, patch client.Patch, opts ...client.SubResourcePatchOption) error {
-					if subResourceName == subResourceStatus {
-						patchCalled = true
-					}
-					return c.SubResource(subResourceName).Patch(ctx, obj, patch, opts...)
-				},
-			})
-
-			controllerReconciler := &MCPServerReconciler{
-				Client: interceptedClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			_, err = controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			Expect(applyCallCount).To(BeNumerically(">", 0), "expected status updates to use SubResourceApply (SSA)")
-			Expect(updateCalled).To(BeFalse(), "status should not use SubResourceUpdate")
-			Expect(patchCalled).To(BeFalse(), "status should not use SubResourcePatch")
-		})
-	})
-
-	Context("When reconciling a resource with health probes", func() {
-		const resourceName = "test-resource-probes"
-
-		ctx := context.Background()
-
-		typeNamespacedName := types.NamespacedName{
-			Name:      resourceName,
-			Namespace: "default",
-		}
-
-		BeforeEach(func() {
-			mcpServer := newTestMCPServer(resourceName)
-			mcpServer.Spec.Runtime.Health = mcpv1alpha1.HealthConfig{
-				LivenessProbe: &corev1.Probe{
-					ProbeHandler: corev1.ProbeHandler{
-						HTTPGet: &corev1.HTTPGetAction{
-							Path: "/health",
-							Port: intstr.FromInt(8080),
-						},
-					},
-					InitialDelaySeconds: 10,
-					PeriodSeconds:       30,
-				},
-				ReadinessProbe: &corev1.Probe{
-					ProbeHandler: corev1.ProbeHandler{
-						TCPSocket: &corev1.TCPSocketAction{
-							Port: intstr.FromInt(8080),
-						},
-					},
-					InitialDelaySeconds: 5,
-					PeriodSeconds:       10,
-				},
-			}
-			Expect(k8sClient.Create(ctx, mcpServer)).To(Succeed())
-		})
-
-		AfterEach(func() {
-			mcpServer := &mcpv1alpha1.MCPServer{}
-			err := k8sClient.Get(ctx, typeNamespacedName, mcpServer)
-			if err == nil {
-				Expect(k8sClient.Delete(ctx, mcpServer)).To(Succeed())
-			}
-		})
-
-		It("should create deployment with health probes", func() {
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers).To(HaveLen(1))
-
-			// Verify liveness probe
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe).NotTo(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.HTTPGet).NotTo(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.HTTPGet.Path).To(Equal("/health"))
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.HTTPGet.Port.IntVal).To(Equal(int32(8080)))
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.InitialDelaySeconds).To(Equal(int32(10)))
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.PeriodSeconds).To(Equal(int32(30)))
-
-			// Verify readiness probe
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe).NotTo(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe.TCPSocket).NotTo(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe.TCPSocket.Port.IntVal).To(Equal(int32(8080)))
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe.InitialDelaySeconds).To(Equal(int32(5)))
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe.PeriodSeconds).To(Equal(int32(10)))
-		})
-
-		It("should update deployment when probes change", func() {
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			By("Reconciling to create the initial deployment with probes")
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.InitialDelaySeconds).To(Equal(int32(10)))
-
-			By("Updating probes")
-			mcpServer := &mcpv1alpha1.MCPServer{}
-			Expect(k8sClient.Get(ctx, typeNamespacedName, mcpServer)).To(Succeed())
-			mcpServer.Spec.Runtime.Health.LivenessProbe = &corev1.Probe{
-				ProbeHandler: corev1.ProbeHandler{
-					HTTPGet: &corev1.HTTPGetAction{
-						Path: "/healthz",
-						Port: intstr.FromInt(8080),
-					},
-				},
-				InitialDelaySeconds: 15,
-				PeriodSeconds:       60,
-			}
-			mcpServer.Spec.Runtime.Health.ReadinessProbe = &corev1.Probe{
-				ProbeHandler: corev1.ProbeHandler{
-					HTTPGet: &corev1.HTTPGetAction{
-						Path: "/ready",
-						Port: intstr.FromInt(8080),
-					},
-				},
-				InitialDelaySeconds: 3,
-				PeriodSeconds:       5,
-			}
-			Expect(k8sClient.Update(ctx, mcpServer)).To(Succeed())
-
-			By("Reconciling again to pick up the change")
-			_, err = controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-
-			// Verify updated liveness probe
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.HTTPGet.Path).To(Equal("/healthz"))
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.InitialDelaySeconds).To(Equal(int32(15)))
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.PeriodSeconds).To(Equal(int32(60)))
-
-			// Verify updated readiness probe
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe.HTTPGet).NotTo(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe.HTTPGet.Path).To(Equal("/ready"))
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe.InitialDelaySeconds).To(Equal(int32(3)))
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe.PeriodSeconds).To(Equal(int32(5)))
-		})
-
-		It("should update deployment when probes are removed", func() {
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			By("Reconciling to create the initial deployment with probes")
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe).NotTo(BeNil())
-
-			By("Removing probes")
-			mcpServer := &mcpv1alpha1.MCPServer{}
-			Expect(k8sClient.Get(ctx, typeNamespacedName, mcpServer)).To(Succeed())
-			mcpServer.Spec.Runtime.Health.LivenessProbe = nil
-			mcpServer.Spec.Runtime.Health.ReadinessProbe = nil
-			Expect(k8sClient.Update(ctx, mcpServer)).To(Succeed())
-
-			By("Reconciling again to pick up the change")
-			_, err = controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      resourceName,
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe).To(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe).To(BeNil())
-		})
-
-		It("should handle only liveness probe (no readiness)", func() {
-			mcpServer := newTestMCPServer("test-only-liveness")
-			mcpServer.Spec.Runtime.Health.LivenessProbe = &corev1.Probe{
-				ProbeHandler: corev1.ProbeHandler{
-					HTTPGet: &corev1.HTTPGetAction{
-						Path: "/health",
-						Port: intstr.FromInt(8080),
-					},
-				},
-				InitialDelaySeconds: 10,
-			}
-			Expect(k8sClient.Create(ctx, mcpServer)).To(Succeed())
-			defer func() {
-				err := k8sClient.Get(ctx, client.ObjectKey{Name: "test-only-liveness", Namespace: "default"}, mcpServer)
-				if err == nil {
-					Expect(k8sClient.Delete(ctx, mcpServer)).To(Succeed())
-				}
-			}()
-
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: "test-only-liveness", Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      "test-only-liveness",
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe).NotTo(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe.HTTPGet).NotTo(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe).To(BeNil())
-		})
-
-		It("should handle only readiness probe (no liveness)", func() {
-			mcpServer := newTestMCPServer("test-only-readiness")
-			mcpServer.Spec.Runtime.Health.ReadinessProbe = &corev1.Probe{
-				ProbeHandler: corev1.ProbeHandler{
-					TCPSocket: &corev1.TCPSocketAction{
-						Port: intstr.FromInt(8080),
-					},
-				},
-				InitialDelaySeconds: 5,
-			}
-			Expect(k8sClient.Create(ctx, mcpServer)).To(Succeed())
-			defer func() {
-				err := k8sClient.Get(ctx, client.ObjectKey{Name: "test-only-readiness", Namespace: "default"}, mcpServer)
-				if err == nil {
-					Expect(k8sClient.Delete(ctx, mcpServer)).To(Succeed())
-				}
-			}()
-
-			controllerReconciler := &MCPServerReconciler{
-				Client: k8sClient,
-				Scheme: k8sClient.Scheme(),
-			}
-
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: "test-only-readiness", Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			deployment := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, client.ObjectKey{
-				Name:      "test-only-readiness",
-				Namespace: "default",
-			}, deployment)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(deployment.Spec.Template.Spec.Containers[0].LivenessProbe).To(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe).NotTo(BeNil())
-			Expect(deployment.Spec.Template.Spec.Containers[0].ReadinessProbe.TCPSocket).NotTo(BeNil())
 		})
 	})
 })
